Fix more warnings, found by cross-compiling Qt
KeccakF-1600-opt32.c:497:5: error: comparison between signed and unsigned integer expressions [-Werror=sign-compare]
harfbuzz-thai.c:264:49: error: comparison is always false due to limited range of data type [-Werror=type-limits]

These warnings are caused by "char" defaulting to unsigned on ARM. In
particular, the second warning was introduced by commit
785e95ef0a95ca8fb39ef57678cd4876ee657c43, which is not upstream...

qbenchmarkvalgrind.cpp:224:5: error: variable ‘_qzz_res’ set but not used [-Werror=unused-but-set-variable]

This one was fixed for x86-64 in 7b54571ec2032628ea71b0af but not
for the other platforms.

KeccakF-1600-opt32.c:250:5: error: comparison between signed and unsigned integer expressions [-Werror=sign-compare]

This one is wasn't caught before because it applies only to big-endian
code.

+++ src/3rdparty/sha3/KeccakF-1600-opt32.c
@@ -114,8 +114,8 @@ static void setInterleavedWordsInto8bytes(UINT8* dest, UINT32 even, UINT32 odd)
 		const UINT32 * pI = (const UINT32 *)input; \
 		UINT32 * pS = state; \
 		UINT32 t, x0, x1; \
-	    int i; \
-	    for (i = (rateInLanes)-1; i >= 0; --i) \
+        int i; \
+        for (i = (int)(rateInLanes)-1; i >= 0; --i) \
 		{ \
 			x0 = *(pI++); \
 			t = (x0 ^ (x0 >>  1)) & 0x22222222UL;  x0 = x0 ^ t ^ (t <<  1); \
@@ -170,7 +170,7 @@ static void xor8bytesIntoInterleavedWords(UINT32* evenAndOdd, const UINT8* sourc
 
 #define xorLanesIntoState(laneCount, state, input) \
     { \
-        int i; \
+        unsigned i; \
         for(i=0; i<(laneCount); i++) \
             xor8bytesIntoInterleavedWords(state+i*2, input+i*8); \
     }
@@ -212,7 +212,7 @@ static void setInterleavedWordsInto8bytes(UINT8* dest, UINT32* evenAndOdd)
 
 #define extractLanes(laneCount, state, data) \
     { \
-        int i; \
+        unsigned i; \
         for(i=0; i<(laneCount); i++) \
             setInterleavedWordsInto8bytes(data+i*8, (UINT32*)state+i*2); \
     }
